ENH: Add upsampled_filter_convolution
Add this for the stationary wavelet transform. Currently only supports
periodization mode, since that is what the old swt does. Since the signal is
intended to remain the same size, I'm not sure how useful signal extension modes
would be. Looks very similar to downsampling_convolution_periodization, but with
the step size referring to the filter step rather than the input step.