ENH: Use stride in downsampling_convolution_periodization
The stride parameter is added to this function only because it is not publicly
exposed, so I can use this for performance and regression testing.

The stride is a naive implementation, with all input[n] replaced by
input[n*stride], and the same for the output.

This could even be implemented as a macro:
`#define ACCESS(array, idx) array[(idx)*stride]`
or when strides are not required:
`#define ACCESS(array, idx) array[idx]`

Futher testing shows that changing loops that modify indexes into input to
modify the index directly by the stride is more difficult than expected, and
results marginal to modest performance decreases, so the naive implementation
is probably best.